种子接口优化1.3

+++ game-resource/src/main/java/com/dkm/seed/entity/vo/UserInIf.java
@@ -25,11 +25,11 @@ public class UserInIf {
     /**
      * 用户当前经验
      */
-    private Long userInfoNowExperience;
+    private Integer userInfoNowExperience;
     /**
      * 用户下一等级所需经验值
      */
-    private Long userInfoNextExperience;
+    private Integer userInfoNextExperience;
     /**
      * 种子等级
      */
